services/sentiment/requirements.txt
feedparser
vaderSentiment
psycopg2-binary
requests

